# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Letpai API
  version: 1.0.0
  description: API for Letpai - Bill splitting application with WhatsApp integration
  contact:
    name: Letpai Team
    email: support@letpai.com
servers:
  - url: http://localhost:3000/api/v1
  - url: /api/v1
paths:
...
  /payments/{participant_id}/submit:
    post:
      summary: Submit payment proof
      description: Submit payment proof (public link, no auth required). Payment submissions are rejected if session has expired (session_date has passed).
      operationId: submitPayment
      tags:
        - Payments
      parameters:
        - $ref: '#/components/parameters/ParticipantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitPaymentRequest'
      responses:
        '201':
          description: Payment proof submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSubmitResponse'
        '400':
          description: Bad request (session expired, payment already submitted, or invalid file)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/{participant_id}/public:
    get:
      summary: Get payment page
      description: Get payment page data (public, no auth required)
      operationId: getPaymentPage
      tags:
        - Payments
      parameters:
        - $ref: '#/components/parameters/ParticipantID'
      responses:
        '200':
          description: Payment page data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentPageResponse'
        '404':
          description: Participant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/{proof_id}/approve:
    post:
      summary: Approve payment
      description: Approve payment (host only)
      operationId: approvePayment
      tags:
        - Payments
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProofID'
      responses:
        '200':
          description: Payment approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentApproveResponse'
  /payments/{proof_id}/reject:
    post:
      summary: Reject payment
      description: Reject payment (host only). A WhatsApp notification will be sent to participant with rejection reason.
      operationId: rejectPayment
      tags:
        - Payments
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProofID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectPaymentRequest'
      responses:
        '200':
          description: Payment rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRejectResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "AUTH_001"
            message:
              type: string
              example: "Invalid credentials"
...
    BillItem:
      type: object
      properties:
        bill_item_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        session_id:
          type: string
          format: uuid
        description:
          type: string
          description: Primary label for the bill item (e.g. "Nasi Goreng")
          example: "Nasi Goreng"
        amount:
          type: number
          format: float
          description: Amount in the session's currency base unit (e.g. IDR rupiah, not cents)
          example: 50000
        category:
          type: string
          nullable: true
          example: "food"
        created_at:
          type: string
          format: date-time
          example: "2026-03-18T12:00:00Z"
        updated_at:
          type: string
          format: date-time
...
    SubmitPaymentRequest:
      type: object
      required:
        - proof_image
        - file_name
      properties:
        proof_image:
          type: string
          description: Base64 encoded image
          example: "data:image/jpeg;base64,/9j/4AAQSkZJRg..."
        file_name:
          type: string
          example: "proof.jpg"
        file_format:
          type: string
          example: "image/jpeg"
    PaymentSubmitResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Payment proof submitted"
        data:
          type: object
          properties:
            proof_id:
              type: string
              format: uuid
            status:
              type: string
              enum: [pending, submitted, paid, rejected]
              example: "submitted"
            uploaded_at:
              type: string
              format: date-time
              example: "2026-03-18T15:00:00Z"
    PaymentPageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            session_name:
              type: string
              example: "Makan Siang"
            participant_name:
              type: string
              example: "John Doe"
            share_amount:
              type: number
              format: float
              example: 125000
            currency:
              type: string
              example: "IDR"
            bill_items:
              type: array
              items:
                $ref: '#/components/schemas/BillItem'
            payment_status:
              type: string
              enum: [pending, submitted, paid, rejected]
              example: "pending"
            payment_proof_url:
              type: string
              format: uri
              nullable: true
            rejection_reason:
              type: string
              nullable: true
            link_expires_at:
              type: string
              format: date-time
              example: "2026-03-25T14:00:00Z"
...
    PaymentApproveResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Payment approved"
        data:
          type: object
          properties:
            payment_status:
              type: string
              example: "paid"
            approved_at:
              type: string
              format: date-time
              example: "2026-03-18T15:30:00Z"
    RejectPaymentRequest:
      type: object
      required:
        - rejection_reason
      properties:
        rejection_reason:
          type: string
          example: "Photo is blurry, please upload clearer proof"
    PaymentRejectResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Payment rejected"
        data:
          type: object
          properties:
            payment_status:
              type: string
              example: "rejected"
            rejection_count:
              type: integer
              example: 1
            rejected_at:
              type: string
              format: date-time
              example: "2026-03-18T15:30:00Z"
...
  parameters:
...
    ParticipantID:
      name: participant_id
      in: path
      required: true
      description: Participant ID
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    ProofID:
      name: proof_id
      in: path
      required: true
      description: Payment Proof ID
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...